Refactor Header
