+++ src/App.vue
@@ -3,7 +3,7 @@ import { onMounted } from 'vue';
 import { RouterView } from 'vue-router';
 
 import RouterTitle from '@/components/layout/RouterTitle.vue';
-import Header from '@/components/layout/Header/Header.vue';
+import Header from '@/components/layout/Header.vue';
 import Footer from '@/components/layout/Footer.vue';
 import UpdateHandlerDialog from '@/components/layout/UpdateHandlerDialog.vue';
 import { useCustomization } from '@/composables/useCustomization';
