--- src/components/layout/Header/Header.vue
@@ -1,11 +1,0 @@
-<script setup>
-import HeaderNav from './HeaderNav.vue';
-import HeaderBrand from './HeaderBrand.vue';
-</script>
-
-<template>
-  <header class="flex flex-col gap-3">
-    <HeaderBrand></HeaderBrand>
-    <HeaderNav></HeaderNav>
-  </header>
-</template>
