--- src/components/layout/Header/HeaderBrand.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="flex h-14 items-center justify-center gap-4">
-    <img alt="" class="size-16 max-w-full object-contain" src="@/assets/logo/round.png" />
-    <div class="flex flex-col text-center">
-      <h1 class="text-2xl">Foxford Tools</h1>
-      <p class="text-xs">кастомизация сайта Фокса</p>
-    </div>
-  </div>
-</template>
