--- src/components/layout/Header/HeaderNav.vue
@@ -1,27 +1,0 @@
-<script setup>
-import { useRouter } from 'vue-router';
-
-import { Button } from '@/components/ui/button';
-
-const router = useRouter();
-const routes = router.getRoutes();
-</script>
-
-<template>
-  <nav>
-    <ul class="flex justify-center gap-3" aria-label="Вкладки">
-      <li v-for="route in routes" :key="route.name">
-        <Button
-          :aria-label="route.meta.title"
-          :class="$route.name === route.name ? 'bg-muted' : ''"
-          size="icon"
-          variant="outline"
-          as-child>
-          <RouterLink :to="{ name: route.name }">
-            <component :is="route.meta.icon" class="text-foreground" stroke-width="1.5" aria-hidden="true" />
-          </RouterLink>
-        </Button>
-      </li>
-    </ul>
-  </nav>
-</template>
